debug: add missing fields, change data types

+++ src/main/java/com/example/storage/Session.java
@@ -13,8 +13,8 @@ public class Session implements Serializable {
     private String sessionId = null;
     // may need min/max timestamp
     //private String timestamp = null;
-    private Float longitude = null;
-    private Float latitude = null;
+    private Double longitude = null;
+    private Double latitude = null;
     private boolean hasTxn = false; // true if the session has at least 1 transaction
     private ArrayList<Action> actions = new ArrayList<Action>(); // TODO: might change datatype later
     
@@ -22,7 +22,7 @@ public class Session implements Serializable {
     public ArrayList<Action> sortedActions() {
         // copy list
         ArrayList<Action> copy = new ArrayList<Action>();
-        copy.addAll(getActions()); // Note: can't modify these actions, they still reference the originals
+        copy.addAll(getActions()); // Note: can't modify the action elements, they still reference the originals
         // sort list
         Collections.sort(copy, (Action a1, Action a2) ->{
             return a1.getTimestamp().compareTo(a2.getTimestamp());
